Revamped the WicketContextExecutorImpl to fix some issues that 01500bda545a4142a20efb6d93bc608978f96f69 did *not* fix
The problem lied around application.getName() in AbstractWebappConfig.wicketContextExecutor: getName() may only be called after the application has been initialized by WicketFilter, which runs only after all beans have been initialized. So application.getName() cannot be used in a @Bean-annotated method: it would result in a null value that would break things later.
The solution was to simply use the application directly, and not its name. That was already the case in WicketContextExecutor anyway: the name was only used to retrieve the application from the Spring application context.

Another problem has been fixed in WicketContextExecutorImpl: retrieveWebApplication used to alter Wicket's ThreadContext, and did so only in some cases. This mechanic has been fixed and moved to attachApplicationIfNeeded.