Removed two overrides that were introduced by mistake
Wicket added a public getChoicesModel() *after* we did so on our own classes,
and I mistakenly added a @Override annotation when discovering the warning.
The effect was that even when setChoices(IModel) was called by users, getChoicesModel()
would still continue to return the original wrapped model. As some methods would use
getChoicesModel(), and some others would use this.choices, this resulted in some
really strange behaviors.

Another problem (to be solved later) is that calling setChoices(IModel) will not wrap
the model.