Roll back to commons-io 2.4
We're stuck at version 2.4 because of an incompatibility with TrueZip from commons-io 2.5 on.

In commit 1721:f836922f4de2 in TrueZip, released in version 7.2, the developer decided *not* to implement
part of the standard Java API, namely File.toPath(), which throws an exception as of this commit.
This decision causes commons-io to fail miserably when performing operations on TFiles, for instance
org.apache.commons.io.FileUtils.deleteDirectory, which calls Java7Support.isSymlink, which calls
TFile.toPath, which throws an exception. Unfortunately, we have no control over this code and cannot decide
not to use File.toPath, so we can't do a thing to solve this issue.

This call to `toPath` in commons-io has been introduced in revision `r1686472` (released in version 2.5).
There was no issue that we know of before.

Rolling back to TrueZip 7.1 is not an option: it's really old and missing features that we're using right
now.
Thus, for now, we'll stick with commons-io 2.4, until we migrate from TrueZip to its successor, TrueVFS,
which seems to have solved the issue.