Don't catch Exception when Exception is not thrown
Catching java.lang.Exception is an extreme measure: it will swallow all sorts of exceptions.

This includes:

 * exceptions that should never have been catched, and should rather have been transfered to the caller, such as (in most cases) InterruptedException.
 * exceptions that will be added in the future to the "throws" clause of called methods (when called methods are updated)

For these reasons, java.lang.Exception should never be catched, unless it is actually thrown as java.lang.Exception (as in Callable, for instance).

This commit turns "catch(Exception)" into "catch(RuntimeException|CheckedException1|CheckedException2|...)" in most places (webapp exceluded). It does not change the way exceptions are handled at all, but only makes the type of exceptions that are swallowed or wrappedd more explicit.