Always handle InterruptedException when catching Exception
If we just swallow the exception, or use it as another exception's cause, then the interruption won't be notified to callers, which is bad practice.
In this commit, the cases where we can't just re-throw the InterruptedException were handled by setting the thread's interrupt status.

See http://www.ibm.com/developerworks/library/j-jtp05236/