hibernate-search 5.9: fix analyzer retrieval
We need consistent behavior to retrieve analyzer for client-side
manipulations both when lucene or elasticsearch are used.

With hsearch 5.9, we cannot use hibernate-search registry to store
lucene analyzers with a special prefix (behavior used with hsearch 5.7).

We now use our own analyzer registry. We'll need to get rid of this as
soon as we can remove all getAnalyzer(fieldName) usages.

An embedded elasticsearch example is provided in ElasticsearchHelper for
future tests or development environment.