infinispan: try to fix infinispan test
To fix TestRestart and TestMessage. Currently, launched processes to
does not wait for cluster initialization before to send message. So
there's clearly a run condition on the possiblity that all messages can
be received by main node.

This modification add an expected view size argument to launched
process, so that task is executed after all nodes are connected.