fix #18: notification groups
Fix on hash() and equals() method allows to group recipient accordingly to
their preferred locale. Before this fix, each recipient of a same
notification call receives a separate email.

Fix on default locale assignment. As locale is coalesced before wicket
rendering (to an available locale, or to a default value if locale is
unset or not available).
Before this fix, two recipients, one with default locale explicitly set,
and the other with a null locale, received separate notification. After
this fix, grouping is done on the effective locale, and mail are correctly
grouped together.