Logging mechanism for PropertySources
This will be used to ensure that property source resolution is not
broken when we migrate from old system to vanilla Spring
@PropertySource.

PropertySourcesPlaceholderConfigurer is a component initialized with
environment that loads properties from @ConfigurationLocations and
bootstrap configuration.

It is needed to dump this configuration in pre-@PropertySource igloo.

Javadoc for ConfigurationLogger.java translated in english.