jackson+tasks: remove deprecated calls
We now use explicit configuration for authorized type for polymorphic
deserialization.

We need to keep a permissive deserialization config, that is acceptable
as data is internally-managed by application.

For console print, we use a node tree instead of a real
deserialization/serialization. It may allow to view content even if
there is a ser/deser issue, and it allows to use a basic ObjectMapper.