Prepare 4.0.0 release
basic-application igloo.version configuration is broken for jgitflow.

When jgitflow encounters a version-as-property, it tries to read
property from local `<properties>` node. If there is no `<properties>`
node, it just ignores this property. If there is a `<properties>` node
not containing the version property, it then fails.

We drop `<properties>` from basic-application-*. This pattern allow
jgitflow to silently ignore this version.

We'll need to find a way to use properties and jgitflow tool.