fix babel - TypeError: class constructors must be invoked with |new|
As we inherit bootstrap object, by default, babel uses a class
transformation pattern (we do not have class ... in transpiled code, and
it triggers an error in browser).

As bootstrap code safely uses class ... syntax for its components, a
better option is to ensure that we do not transform classes. This is
done by this configuration change.