bring back modal-more isTransitioning workaround
When modal is in transition state, it is not possible de toggle its
state. This is a problem for work-in-progress modal as it closing is
triggered by file download termination, and for small file it is short
enough so that modal stay opened.

We reuse the bootstrap 4 workaround where transitioning state is
enforced to false when show/hide is called, so show/hide is always
honored.