igloo-boot: workaround to force incrementSize to 1
With Hibernate 6, default increment size for sequence is 50. It is done
to allow the use of a PooledOptimizer to generated identifiers (to allow
better performance on batch insert).

If you want a classic 1-increment on sequences, it is needed to declare
your increment size on each @GeneratedValue.

This NoopIncrementOptimizer can be configured with
`hibernate.id.optimizer.pooled.preferred=org.iglooproject.jpa.hibernate.jpa.NoopIncrementOptimizer`
so that all sequences are configured with an increment of 1 and not
optimization. Beware that it disables PooledOptimizer usage.