Use a unique timestamp for CSS #84
Previous implementation uses Long.MAX_VALUE. As lastModifiedDate is used
by wicket to generate a filename suffix to ensure that resource is
refreshed by navigator when needed, we must use a dynamic value.

We use current date for lastModifiedDate. It is cached by application and
reused for each URL generation, so that browser cache can work as
expected. If cache is emptied (application restart), then a new date is
used and browser refresh the resource.